+++ packages/ts-newskit/src/slices/content-bucket-3/index.tsx
@@ -103,6 +103,11 @@ export const ContentBucket3 = ({
                 article={{
                   ...leadArticleRight,
                   hasTopBorder: isMobile,
+                  topBorderStyle: 'lightDashedDivider',
+                  tagAndFlagMarginBlockStart: {
+                    xs: 'space030',
+                    md: 'space040'
+                  },
                   isLeadImage: true,
                   titleTypographyPreset: isLarge
                     ? 'editorialHeadline030'
@@ -131,13 +136,16 @@ export const ContentBucket3 = ({
           <Divider
             overrides={{
               marginBlock: 'space040',
-              stylePreset: 'dashedDivider'
+              stylePreset: { xs: 'lightDashedDivider', md: 'dashedDivider' }
             }}
           />
         </FullWidthHidden>
         <BlockItem>
           <ArticleStack
-            articles={articles}
+            articles={articles.map(article => ({
+              ...article,
+              topBorderStyle: { xs: 'lightDashedDivider', md: 'dashedDivider' }
+            }))}
             breakpoint={currentBreakpoint}
             clickHandler={clickHandler}
           />