+++ packages/article-comments/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.40.38](https://github.com/newsuk/times-components/compare/@times-components/article-comments@0.40.37...@times-components/article-comments@0.40.38) (2023-10-19)
+
+**Note:** Version bump only for package @times-components/article-comments
+
+
+
+
+
 ## [0.40.37](https://github.com/newsuk/times-components/compare/@times-components/article-comments@0.40.36...@times-components/article-comments@0.40.37) (2023-10-18)
 
 **Note:** Version bump only for package @times-components/article-comments