+++ packages/star-button/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.43](https://github.com/newsuk/times-components/compare/@times-components/star-button@0.4.42...@times-components/star-button@0.4.43) (2023-10-19)
+
+**Note:** Version bump only for package @times-components/star-button
+
+
+
+
+
 ## [0.4.42](https://github.com/newsuk/times-components/compare/@times-components/star-button@0.4.41...@times-components/star-button@0.4.42) (2023-10-18)
 
 **Note:** Version bump only for package @times-components/star-button