+++ packages/utils/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.16.21](https://github.com/newsuk/times-components/compare/@times-components/utils@6.16.20...@times-components/utils@6.16.21) (2023-10-19)
+
+**Note:** Version bump only for package @times-components/utils
+
+
+
+
+
 ## [6.16.20](https://github.com/newsuk/times-components/compare/@times-components/utils@6.16.19...@times-components/utils@6.16.20) (2023-10-18)
 
 **Note:** Version bump only for package @times-components/utils